Add improved browser detection
The goal is to move away from browser detection for features. Specific browser will be used only for things like advertisement or special rendering.

Feature differentiation should be based on API availability as much as possible instead.

This could create problems if someone tries to manually install the wrong version of the extension or if they install it from the wrong store (Ex: install the extension from chrome web store on opera).